Move `UpdateComment` command
