+++ wcfsetup/install/files/lib/command/comment/UpdateComment.class.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace wcf\system\comment\command;
+namespace wcf\command\comment;
 
 use wcf\data\comment\Comment;
 use wcf\data\comment\CommentAction;
@@ -15,15 +15,14 @@ use wcf\system\message\embedded\object\MessageEmbeddedObjectManager;
  * @author      Marcel Werk
  * @copyright   2001-2024 WoltLab GmbH
  * @license     GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
- * @since       6.1
+ * @since       6.2
  */
 final class UpdateComment
 {
     public function __construct(
         private readonly Comment $comment,
         private readonly HtmlInputProcessor $htmlInputProcessor,
-    ) {
-    }
+    ) {}
 
     public function __invoke(): void
     {
